共用tasker界面
将默认tasker.interface()继承给default类tasker，extra类tasker
